is-extraneous: Rewrite to use the logical tree
The basic rule is:

If this package is disconnected from the top of the tree then it is
extraneous.

With the following exceptions:

If this package was (or is connected to a package that was) required by a
user directly, eg `npm install foo`, then we don't consider it extraneous.
That is, the summary tree from `npm install foo` shouldn't show the newly
installed modules as extraneous, though `npm ls` will show them as such.

If there is no top level package.json then nothing is extraneous.