npm-cli: Eliminate output corruption
On Windows stdout & stderr are always blocking.  Prior to Node 6, this was
true on Mac too. This has never been true on Linux.

When they're non-blocking it's possible to end up with two undesirable
things happening:

First a buffer can flush part way through a line, allowing output from
stderr to show up on the same line as output from stdout (or vice-versa).
This results in at best, nonsensical mixed text.  But it's worse if you're
using unicode characters.  Unicode characters aren't atomic, so it's
possible for a buffer to be split mid-character, resulting in invalid
characters in the output tty.

Second, we output our progress bar to stderr. Before printing things
to stdout we want to erase the progress bar so they don't overlap. So
we need to be able to wait for the erase to be flushed before printing
anything to stdout, but there's no way to do that.

Making the streams blocking solves all of these problems.  It's worth noting
that `setBlocking` only applies to tty streams.  So pipes and redirections
will continue to be async.